This is a changed version of the engine3d website
The website includes auto generation of doxygen xml which is then used by breath and sphinx. Breath hierarchy is also auto generated using doxygen xml.

New side bar to showcase heirarchy and to integrate sphinx with mkdocs.

What needs to be done, eventually. A tab should be built for each function that seperates the parameters, from our reasoning and examples.

Title also needs to be fixed. Overall it needs to look better lol. Not my type of skill set though.